Release 0.7.13: Support for doxygen versions for 1.8.13 to 1.8.20
- Add self.options to track detected doxygen versions
- Rework whitespace handling
- Enable and disable features depending on what version of doxygen was used.

A lot of this can be inferred by what is seen within the xml content
generated. Older versions of doxygen do not support latexonly so we
automatically fall back to xmlonly material that is embedded in <para>
blocks.  If we detect latexonly or htmlonly blocks, we are almost
guaranteed that we can ignore any xmlonly content.